Yeni klasör/backend/pyproject.toml
[project]
name = "eduai-backend"
version = "1.0.0"
description = "Backend for EduAI - Intelligent Learning Platform"
authors = [
    {name = "EduAI Team"}
]
dependencies = [
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "sqlalchemy>=1.4.0",
    "pydantic-settings>=2.0.0",
    "passlib>=1.7.4",
    "bcrypt>=3.2.0",
    "python-multipart>=0.0.5",
    "python-jose[cryptography]>=3.3.0",
    "python-dotenv>=0.19.0",
    "email-validator>=1.1.3",
    "langchain>=0.1.0",
    "langchain-google-genai>=0.1.0",
    "langchain-tavily>=0.1.0",
    "tavily-python>=0.3.0",
    "google-generativeai>=0.3.0",
    "aiohttp>=3.8.0",
    "crawl4ai>=0.2.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "playwright>=1.54.0",
    "firecrawl-py>=2.16.2",
    "requests>=2.32.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.2.4",
    "black>=21.7b0",
    "flake8>=3.9.2"
]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"
